A lot has been changed and added.
We're currently in the process of working towards an official login screen that has proper authentication.

I started work on an admin's only screen, but that has been put on hold until we build the rest of the foundation.

The code is going to need some major revamp to ensure that we aren't facing much technical debt in the future.

The next commit from me should hopefully contain a fully working login screen and match display list on the home screen.