Added timer periodic, driver station dropdown, and cleaned up some stuff..
Added a timer periodic to send cached match data to the server.

Added a little dropdown in the match form to allow users to hand select the alliance color and driver number.

Temporarily removed the list view for the assigned and all matches. That's not implemented server-side and we don't have the feature really ready.

Login should be done and do what we expect. Tag should be the one to finish the details regarding it though. (located in pages/login_as_user.dart).

When testing make sure to put your server host address into the globals.dart file.