Got the stuff somewhat working for bluetooth!
I can't believe it! It's not CRASHING!!!! WOOOOOOOOO!!!!!

Finally, the android build has some of the fixes it needs. For some reason, the bluetooth advertise permission wasn't enabled so I had to force it with a permission request.

I now need to test this with 3 to 4 devices all connecting and disconnecting when sending data. I also need to make the sample data more "unique" so it's easier to tell who sends the data and be able to visually separate it.