Worked on improving the set up of data for home.
I needed to reroute the manner for which we get the data about "assigned" and "all" matches. I wanted to move that over to a easy singleton object to call.

I also made a wrapper for http post and http get (get is still in progress). This should make further requests to the server a breeze.

The UI of the home page is not final and I feel could be done a bit better, but for now I will stick to this simplistic look.

Once I have an actual server to test against (either on tag's server or locally running it on my machine), then I can make sure that the pulled data will get formatted correctly into display info on the templated forms.