Updated the info we collect for matches.
Tag brought up a good point in that we should probably quickly get the new data that we would be needing for the matches up ahead.

I ended up updating our example json and match_form page. I also was doing work on Admin, but that isn't finished yet.

But, let me quickly tell you what I plan for admin. Admin is going to have a dedicated page for assigning matches to users. This menu will have a dropdown for every user available and a set of input fields to specify assigning a range of matches with a specific color and drive team number.

Hopefully I can implement this by tonight.