Worked on a separate group assign matches page.
We will have two pages, individual and group assignment pages for admins. This allows admins to have the option for both granularity and quick easy tasks.

I hope that this system will be fully fleshed out by tomorrow. I'm stopping work for tonight.

Full integration should be done by tomorrow hopefully and so should the admin stuff. Then we can work on cosmetic changes that help give the app a polished and user-friendly look and feel.