Added internet indicator.
We can now teel whether or not we are currently connected to the server or internet in general.

I've also made a placeholder for the leaderboard menu.

I've also switched the periodic from being a post request to a socket connect in main. Hopefully this will be less intensive because it isn't asking the server for anything other than to check if it can connect with it.