Debug Info Implemented.
You can now reset caches. You can force saved caches to be uploaded.

Also fixed an issue for the assigned matches didn't actually assign the match number and team number to the json.

One annoyance for me is the fact that the matches don't snuggly fit into the bottom of the page, but I can't be bothered for now.

I also attempted to add qr scanner support, but it's going to need quite a bit of work to find a proper library for this.