Swapped around the place for Save/
There is currently a bug on safari where the website can suddenly get really messed up. To combat this, we want to make the saving of data easier to access. In the future, this will be helpful because hopefully we can just pull saved matches and "edit" them.

For now, the action of saving is one time, but by implementing this we will future proof it for future user accessibility.

Also, removed the qr scanner button on home. It's because we don't actually have qr support yet.