We now have mangled data support.
We can now display data that has been mangled or just isn't appropriate to be parsed in the all time cache.

It doesn't allow users to resend this mangled data, but we now support copying and pasting the data. This should help with debugging in the future.

Also, added proper checks to make sure we aren't just responding to every response as successful.