Finished implementing leaderboard.
The leaderboard now successfully pulls from the server.

Also, I changed the individual and group match assigns for admins to use
a stream builder so that it can re-build as soon as it gets all the
users, which sometimes can take a bit. This should improve the
experience and not require admins to exit the page and re-enter to get
the updated list.

Stream Builder's are actually nice...