New Match Form Page! (Not Finished)
I started work on a new layout for the match form page! It's supposed to
be optimized for the type of data collection we do and to make it much
more convenient for scouters (Hopefully).

So far it's turning out well and I hope to get the frontend done by
tommorow.