Added rescout button to the bottom of the match form page.
I didn't want to wind up and add a bunch of new things for rescouting.
We have all the data there already in a nice form. I just warn the user
that the flag they're setting is irreversible and that they shouldn't
probably touch unless they know what they're doing.

I also fixed a behavior with the floating toggle button. Turns out I
wasn't properly capturing it's state with onPress actions. I decided
that it was probably better to force it to take in a value and to assign
to that instead. Because oftentimes you want the value of toggle in the
upper level anyways.

I don't think I have anything else to report.