Switched Around App Data Storage
Instead of storing everything in a global file with a bunch of loose fitting functions, we can just use getter and setter variables. All of this can also just be stored in one singular singleton class.

It should make future organization easier, especially considering it puts it all in one class that makes it easier to find.