Added documentation and moved things around.
I've added some inline documentation, cleaned up unused code, fixed some terrible formatting, and moved stuff around.

Oh, I also removed the qr scanner and stuff related to that because none of it works so might as well remove it. Though, a lot of the scanner logic actually worked with one of the plugins but wouldn't work when deployed.

I created a general utils file which can continue a bunch of things that may be useful in the future.

I feel like that at this point, I most likely can start writing proper documentation that lives outside of the code. That includes with improving the main readme.

Overall, this commit is going to contain a **lot** of changes.