Fixed the edit users page crashing and customized it.
Now, you have the option to select badges that will be applied to users.

I've also made a change that simplifies how we do http requests. Since a lot of the code between the functions were the same, I thought it would be best to just make it one function.

Additionally, I made some changes to the achievements page to try and make it scale better on smaller devices.

That's all for now.