Felt inspired to document more.
I looked at what Tag had documented for the backend, I felt that what I left behind in the code in terms of documentation was lacking.

So, I wanted to just do a mass documentation update and clean up of some of the code. That way when somebody picks up this project, they can easily start reading the code to understand the project. At least I hope so.

Also, for any future developer that reads this message, I wish you the best of luck with maintaining this repo! It's been a pet project of mine for several months and I hope that it can bloom into something even greater.